data-model: moved checkEquivalence to immutable grant
