+++ packages/data-model/src/immutable/data-grant.ts
@@ -1,7 +1,8 @@
 import { DataFactory } from 'n3';
 import { NamedNode } from '@rdfjs/types';
 import { INTEROP, RDF } from '@janeirodigital/interop-namespaces';
-import { ApplicationFactory, AuthorizationAgentFactory, ImmutableResource } from '..';
+import { ApplicationFactory, AuthorizationAgentFactory, DataGrant, ImmutableResource } from '..';
+import { getAllMatchingQuads, getOneMatchingQuad } from '@janeirodigital/interop-utils';
 
 type StringData = {
   dataOwner: string;
@@ -59,5 +60,85 @@ export class ImmutableDataGrant extends ImmutableResource {
         this.dataset.add(DataFactory.quad(DataFactory.namedNode(child), INTEROP.inheritsFromGrant, thisNode));
       }
     }
+  }
+  public checkEquivalence(otherGrant: DataGrant): boolean {
+    const predicates = [
+      INTEROP.dataOwner,
+      INTEROP.registeredShapeTree,
+      INTEROP.hasDataRegistration,
+      INTEROP.scopeOfGrant
+    ];
+    // INTEROP.inheritsFromGrant
+    // we don't check values, just if both exist or both don't exist
+    const generatedInherits = getAllMatchingQuads(
+      otherGrant.dataset,
+      DataFactory.namedNode(otherGrant.iri),
+      INTEROP.inheritsFromGrant
+    );
+    const equivalentInherits = getAllMatchingQuads(
+      this.dataset,
+      DataFactory.namedNode(this.iri),
+      INTEROP.inheritsFromGrant
+    );
+    if (generatedInherits.length !== equivalentInherits.length) return false;
+
+    // INTEROP.inheritsFromGrant - inverse
+    // we only check if number of statements is the same
+    const generatedInverseInherits = getAllMatchingQuads(
+      otherGrant.dataset,
+      null,
+      INTEROP.inheritsFromGrant,
+      DataFactory.namedNode(otherGrant.iri)
+    );
+    const equivalentInverseInherits = getAllMatchingQuads(
+      this.dataset,
+      null,
+      INTEROP.inheritsFromGrant,
+      DataFactory.namedNode(this.iri)
+    );
+    // TODO(samurex): missing coverage
+    if (generatedInverseInherits.length !== equivalentInverseInherits.length) return false;
+
+    // INTEROP.delegationOfGrant
+    // we check if either both don't exist or both exist
+    // if both exist we compare values
+    const generatedDelegation = getOneMatchingQuad(
+      otherGrant.dataset,
+      DataFactory.namedNode(otherGrant.iri),
+      INTEROP.delegationOfGrant
+    )?.object;
+    const equivalentDelegation = getOneMatchingQuad(
+      this.dataset,
+      DataFactory.namedNode(this.iri),
+      INTEROP.delegationOfGrant
+    )?.object;
+    if (generatedDelegation?.value !== equivalentDelegation?.value) return false;
+
+    for (const predicate of predicates) {
+      const generatedObject = getOneMatchingQuad(
+        otherGrant.dataset,
+        DataFactory.namedNode(otherGrant.iri),
+        predicate
+      ).object;
+      const equivalentObject = getOneMatchingQuad(this.dataset, DataFactory.namedNode(this.iri), predicate).object;
+      if (generatedObject.value !== equivalentObject.value) return false;
+    }
+
+    // INTEROP.accessMode
+    const generatedAccessModes = getAllMatchingQuads(
+      otherGrant.dataset,
+      DataFactory.namedNode(otherGrant.iri),
+      INTEROP.accessMode
+    ).map((quad) => quad.object.value);
+    const equivalentAccessModes = getAllMatchingQuads(
+      this.dataset,
+      DataFactory.namedNode(this.iri),
+      INTEROP.accessMode
+    ).map((quad) => quad.object.value);
+
+    return (
+      generatedAccessModes.length === equivalentAccessModes.length &&
+      generatedAccessModes.every((mode) => equivalentAccessModes.includes(mode))
+    );
   }
 }
